requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlite3-python==1.0.0
colorama==0.4.6
